contrib/blender/python/requirements.txt
# Operational requirements:
phabricator
pyyaml
python-dotenv
python-dateutil
python-slugify
Unidecode

# Development requirements:
black
mypy
pytest
